Fix decode, remove unwanted files
