--- egs/fisher_swbd/ASR/streaming_conformer_ctc/asr_datamodule.py
@@ -1,1 +1,0 @@
-../conformer_ctc/asr_datamodule.py
