--- egs/fisher_swbd/ASR/streaming_conformer_ctc/label_smoothing.py
@@ -1,1 +1,0 @@
-../conformer_ctc/label_smoothing.py
