--- egs/fisher_swbd/ASR/streaming_conformer_ctc/subsampling.py
@@ -1,1 +1,0 @@
-../conformer_ctc/subsampling.py
