--- egs/fisher_swbd/ASR/tdnn_lstm_ctc/README.md
@@ -1,4 +1,0 @@
-
-Please visit
-<https://icefall.readthedocs.io/en/latest/recipes/librispeech/tdnn_lstm_ctc.html>
-for how to run this recipe.
